теперь в Program только Main + не выводятся сожжёные калории

+++ Core/Models/User.cs
@@ -3,13 +3,12 @@ namespace Дневник_Питания.Core.Models
 {
     public class User
     {
-        public int Height { get; set; }
-        public int Weight { get; set; }
-        public int Age { get; set; }
-        public string Gender { get; set; }
-        public string ActivityLevel { get; set; }
-        public double BMR { get; set; }
-        public double TargetCalories { get; set; }
-        
+        public int Height { get; set; } // Рост в см
+        public double Weight { get; set; } // Вес в кг
+        public int Age { get; set; } // Возраст
+        public string Gender { get; set; } // Пол ("м" или "ж")
+        public string ActivityLevel { get; set; } // Уровень активности
+        public double BMR { get; set; } // Основной обмен веществ
+        public double TargetCalories { get; set; } // Целевая калорийность (например, для похудения)
     }
 }
